+++ generate/config.go
@@ -22,20 +22,21 @@ type Config struct {
 	// The following fields are documented in the [genqlient.yaml docs].
 	//
 	// [genqlient.yaml docs]: https://github.com/Khan/genqlient/blob/main/docs/genqlient.yaml
-	Schema              StringList              `yaml:"schema"`
-	Operations          StringList              `yaml:"operations"`
-	Generated           string                  `yaml:"generated"`
-	Package             string                  `yaml:"package"`
-	ExportOperations    string                  `yaml:"export_operations"`
-	ContextType         string                  `yaml:"context_type"`
-	ClientGetter        string                  `yaml:"client_getter"`
-	Bindings            map[string]*TypeBinding `yaml:"bindings"`
-	PackageBindings     []*PackageBinding       `yaml:"package_bindings"`
-	Casing              Casing                  `yaml:"casing"`
-	Optional            string                  `yaml:"optional"`
-	OptionalGenericType string                  `yaml:"optional_generic_type"`
-	StructReferences    bool                    `yaml:"use_struct_references"`
-	Extensions          bool                    `yaml:"use_extensions"`
+	Schema                          StringList              `yaml:"schema"`
+	Operations                      StringList              `yaml:"operations"`
+	Generated                       string                  `yaml:"generated"`
+	Package                         string                  `yaml:"package"`
+	ExportOperations                string                  `yaml:"export_operations"`
+	ContextType                     string                  `yaml:"context_type"`
+	ClientGetter                    string                  `yaml:"client_getter"`
+	Bindings                        map[string]*TypeBinding `yaml:"bindings"`
+	PackageBindings                 []*PackageBinding       `yaml:"package_bindings"`
+	Casing                          Casing                  `yaml:"casing"`
+	Optional                        string                  `yaml:"optional"`
+	OptionalGenericType             string                  `yaml:"optional_generic_type"`
+	StructReferences                bool                    `yaml:"use_struct_references"`
+	Extensions                      bool                    `yaml:"use_extensions"`
+	OmitUnreferencedImplementations bool                    `yaml:"omit_unreferenced_implementations"`
 
 	// The directory of the config-file (relative to which all the other paths
 	// are resolved).  Set by ValidateAndFillDefaults.